use long instead of int to accomadate lager values

+++ modules/adb/src/org/apache/axis2/databinding/types/Time.java
@@ -104,7 +104,7 @@ public class Time implements java.io.Serializable {
         int hour = 0;
         int minite = 0;
         int second = 0;
-        int miliSecond = 0;
+        long miliSecond = 0;
 
         int timeZoneOffSet = TimeZone.getDefault().getRawOffset();
          int milliSecondPartLength = 0;
@@ -195,7 +195,7 @@ public class Time implements java.io.Serializable {
                 miliSecond = miliSecond / 10;
             }
         }
-        calendar.set(Calendar.MILLISECOND, miliSecond);
+        calendar.set(Calendar.MILLISECOND, (int)miliSecond);
         calendar.set(Calendar.ZONE_OFFSET, timeZoneOffSet);
         calendar.set(Calendar.DST_OFFSET, 0);
